lab-02/requirements.txt
openai==1.51.0
streamlit==1.39.0
python-dotenv==1.0.1
requests
autogen==0.3.1
autogenstudio==0.1.5
fpdf
pillow